move hooks out of utils
